adjust Primary and Timeserver for DER times
Primary and Timeserver in both reference implementation and demo now have
functionality to distribute and process timeserver attestations in DER.
For the XMLRPC call in the demo, a wrapper function and Binary object
are employed. The demo is still mangling the data in the XMLRPC call
and will require a bit more experimentation.