Update sample DER metadata files
After the recent change to the way that DER files are signed,
new samples have been generated.

DER data used to be signed by taking its hash, wrapping that hash in
quotation marks (oversight) in a string and encoding that hash in utf-8,
then signing that.

That is unnecessary and complicates things for the client, so now,
DER data is signed by taking its hash, then signing that hash in raw
binary format. This applies to all .der samples.

Additionally, the timeserver attestations included now contain the
same data, in .json and in .der.