DOC: Revamp end of demo: recovery and final description
Instead of exiting the servers, we repair in place, doing only what
is necessary (key revocations, replacement of images).

Also added description about what's happening and why.

Also removed the last attack, which seemed to claim to demonstrate
what it was not demonstrating: it seemed to suggest that it was
showing that the revoked keys were no longer trusted, but the
attack did not employ those keys. A further attack should be added
that does. Issue #63 has been added to document this.