Fix bug in vehicle manifest ASN.1 conversion nesting
In some circumstances, processing a Vehicle Manifest with
multiple ECU manifests in it fails because on the first
pass, the temp variable containing ECU Manifests was
corrupted for the second pass by what was intended to be
a second temp variable assignment that had the same name.
This second variable was unnecessary and has been removed.