Add redundant Image Repo sample files and update zip archives
of sample metadata with the new, long-expiry sample data.

The redundancy is this: there are two folders of sample data,
initial_w_no_update and update_to_one_ecu. It happens to be
the case that the image repo metadata would be the same in
both cases (as assigning an update to an ECU doesn't change
the image repo). Rather than just excluding the duplicate
sample files, I include them now for completeness, so that
each sample set stands on its own.