Primary will no longer always provide JSON time attestation
The Primary correctly received and validated time attestations
in the format configured in tuf.conf.METADATA_FORMAT.

However, the Primary converts time attestations into a
comprehensible JSON-compatible format after receiving and
validating them.

Due to an oversight, the Primary was failing to check the
format setting and (if appropriate) convert time attestations
back into ASN.1/DER before passing them on to the Secondary,
so regardless of the setting and regardless of the format
that Timeserver sent to Primary, Secondaries were always
receiving time attestations in JSON format.

This is now fixed. Accompanying this are the changes
necessary to the demo to allow it to now transfer this
binary data appropriately (wrapping in xmlrpc Binary objects
based on format setting).