Fix file not closed warnings
I ran into warnings about files not being closed at these lines,
which used to say json.load(open(<fname>, 'r')).

At first, I thought that would have resulted in the file being
close when the variable left scope, but, in retrospect, the file
object was returned to json.load, which would not have closed it
on its own.

This fixes that by manually opening and closing the file.
Is this the best Pythonic way to do this?