Remove demo hack in inventorydb, raising SpoofingError correctly
The external web frontend used to re-register an ECU every time
it selected a new target file for it, which was unnecessary. I've
fixed that, and so now inventorydb.register_ecu can correctly
raise uptane.Spoofing if there is an attempt to re-register an
already registered ECU (without it being explicitly expected to
overwrite an existing registration). This also updates the test
in test_director to expect the correct behavior.

Also spaces out a few lines that Vlad found hard to eyeball
due to cramping, and clarifies a comment.