Fix new tests in test_primary by using offline data
in the same way that test_secondary does: importing existing
pinned.json file and adjusting it in-memory to point to the temp
files that contain the test data to load.

Also adds TODO to the original test_secondary and copied
test_primary code reminding me to address the different-OS question
for file:// links. Also adds a bit of space for readability in
one section.