ASN1: Remove BitStrings and clean up ASN.1/DER conversion
Because older code tried to support both BitStrings and OctetStrings,
there was an awkward additional layer of data structures in the ASN.1
definitions, class BinaryData, which could be occupied by either
BitStrings or OctetStrings. BitStrings and BinaryData have been
removed, and anything that is intended to be binary data is now just
a straight OctetString. BitStrings were being encoded as strings
anyway (as opposed to the OctetString, which is encoded as straight
bytes).

In the process, I also simplified the tag constraints on affected
fields. More simplification of other fields will follow in future
commits.