Fix sample metadata archives: remove breaking additional folder level
This fixes test_secondary on this branch as a result (and presumably
a few tests in test_director).

The full_metadata_archive.zip samples were expanding into a directory
containing 'director' and 'imagerepo' directories, instead of simply
producing 'director' and 'imagerepo' directories in place. This was
due to the way that the samples archives were produced: manually on
OS X.