Consolidate and update sample/test ECU Manifests (after ASN1 changes)
There used to be two sets of good ECU Manifests for testing and as samples
for sample ECU 'TCUdemocar'. The duplicate in dir flawed_manifests (which
was misleading anyway) has been removed, the samples in dir samples/ have
been updated to match the simplified ASN1 definitions, and the test that
used the ones in dir flawed_manifests has been updated to use the ones
in samples/.

Additionally, all the DER sample flawed ECU Manifests for testing have
been re-encoded with the new ASN1 simplifications.