DEMO: Fix XMLRPC bug in Pri<->Sec communication in Python2.7.14
Fixes Issue # 150.

When in DER mode (default), the function exposed in the demo
Primary's XMLRPC interface to return the latest timeserver
attestation for use by the demo Secondary was not correctly
wrapping the DER encoded attestation in an xmlrpc Binary()
object for transmission. This broke the demo Secondary's
ability to receive the timeserver attestation in Python
2.7.14, which came out about two months ago and now ships
with OS X High Sierra (and some Linux distros).

In particular, the direct exposure of a function from the
reference implementation that has no concept of XMLRPC
was swapped out of the XMLRPC interface in favor of a
higher level function in the demo Primary module.