DEMO: Remember to save new firmware the compromise example, too
There was a minor bug at the end of the demo attacks, where the
ECU wasn't "installing" the compromised firmware it validated
(vouched for by both repos). The result was that after the
compromise was resolved, the ECU was acting as if it already had
the new, safe firmware at the very end of the demo, which would be
confusing.

I moved the compromised banner check a bit to deal with this
without disrupting control flow.