Fix unencountered bug if multiple Primaries and optional arg
If multiple Primaries are initialized in the same process and
both are initialized without specifying a list of their
Secondaries, then they will both point to the same list of
Secondaries, updating together, rather than having distinct
lists. This is a Python quirk related to the use of mutables
as default arguments in a function. Mutables should not be
default values because they become common across all calls
made to that function. (It is unlikely that this would have
occurred for anyone, unlikely that any implementer would
inherit this issue in another language, and unlikely to run
multiple Primaries in the same process regardless; however,
it is best to avoid this anti-pattern in Python.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>