Use constants to represent datatypes (not raw strings)
Time Attestations, ECU Manifests, and Vehicle Manifests are the
Uptane-unique data types that can be converted to and from
ASN.1/DER representation. Rather than using 'ecu_manifest' or
the like, we'll use constants defined in
uptane.encoding.asn1_codec from now on. (DATATYPE_ECU_MANIFEST
etc.). This commit adds those constants, imports them in
affected modules, and uses them in place of the raw string
descriptions previously used as types. In calls, the
keyword for the argument is no longer used, which required
some arguments in function calls to be re-ordered in a few
cases.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>