Demo: clean up after single-command demo run
in a manner that will work for both interactive Python sessions
and script runs (atexit).

Deletes director/ and imagerepo/ as well as firmware.img, which
are created when the demo is run.

Also removes readline/rlcompleter imports and use, since the
script may not be run as an interactive session.

Also removes an old commented-out import.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>