Correct ASN.1/DER samples after fix to root asn1 conversion
In order for the tests to run correctly, some sample files
have to be updated. This commit provides new sample data
now that the following bug has been fixed:

Bug caused ASN.1 representations of the targets key public
value listed in root metadata to be encoded correctly.

The bug itself was introduced in:
https://github.com/awwad/tuf/pull/10/commits/38b70984de7a8fc6bb7ffe433ce8e64548862687
(If commit hash changes due to rebasing, note that the commit
summary was "ASN1: Remove BitStrings and clean up ASN.1/DER
conversion")

The bug itself was fixed in:
https://github.com/awwad/tuf/pull/10/commits/dc558a2c862e53442bd25e9e397376f5d126699f
(If the commit hash changes due to rebasing, note that the commit
summary was "Fix bug in preceding commit in root_asn1_coder
for key conversion")

The only files that had to change as a direct consequence were the
root.der files. Unfortunately, this meant that the snapshot.der
files had to change, since in Uptane's current version of TUF,
the snapshot role lists the hash and length of the root file.
Consequently, timestamp had to change, since that has the hash
and length of the snapshot file. So I generated wholly new der
sample metadata. These are also updated in the zip archives that
the tests use.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>