Make clients check Director first
and also make the primary and secondary top-level metadata refresh
procedure identical (for future modularization).  This includes a
rename in primary.py of refresh_toplevel_metadata_from_repositories
to just refresh_toplevel_metadata, which is the same name the
function has in secondary.py.  While both names are accurate in a
sense, the former could be misleading in secondary.py.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>